+++ src/DictionaryTest.php
@@ -31,7 +31,7 @@ final class DictionaryTest extends StructuredFieldTest
 
         self::assertEquals($arrayParams, iterator_to_array($instance));
         $instance->clear();
-        self::assertTrue($instance->isEmpty());
+        self::assertTrue($instance->hasNoMembers());
         self::assertFalse($instance->hasMembers());
     }
 
@@ -63,7 +63,7 @@ final class DictionaryTest extends StructuredFieldTest
         $instance = Dictionary::fromAssociative($arrayParams);
 
         self::assertCount(2, $instance);
-        self::assertFalse($instance->isEmpty());
+        self::assertFalse($instance->hasNoMembers());
         $instance->delete('boolean');
 
         self::assertCount(1, $instance);
@@ -82,7 +82,7 @@ final class DictionaryTest extends StructuredFieldTest
 
         $instance->delete('foobar', 'string');
         self::assertCount(0, $instance);
-        self::assertTrue($instance->isEmpty());
+        self::assertTrue($instance->hasNoMembers());
     }
 
     /** @test */