+++ src/InnerListTest.php
@@ -19,14 +19,14 @@ final class InnerListTest extends TestCase
         $booleanItem = Item::from(true);
         $arrayParams = [$stringItem, $booleanItem];
         $instance = InnerList::fromList($arrayParams, Parameters::fromAssociative(['test' => Item::from(42)]));
-        self::assertFalse($instance->parameters->isEmpty());
+        self::assertFalse($instance->parameters->hasNoMembers());
 
         self::assertSame($stringItem, $instance->get(0));
-        self::assertFalse($instance->isEmpty());
+        self::assertFalse($instance->hasNoMembers());
 
         self::assertEquals($arrayParams, iterator_to_array($instance));
         $instance->clear();
-        self::assertTrue($instance->isEmpty());
+        self::assertTrue($instance->hasNoMembers());
         self::assertFalse($instance->hasMembers());
     }
 
@@ -40,7 +40,7 @@ final class InnerListTest extends TestCase
 
         self::assertCount(2, $instance);
         self::assertTrue($instance->has(1));
-        self::assertTrue($instance->parameters->isEmpty());
+        self::assertTrue($instance->parameters->hasNoMembers());
         self::assertFalse($instance->parameters->hasMembers());
 
         $instance->remove(1);
@@ -57,7 +57,7 @@ final class InnerListTest extends TestCase
 
         $instance->remove(0, 1);
         self::assertCount(0, $instance);
-        self::assertTrue($instance->isEmpty());
+        self::assertTrue($instance->hasNoMembers());
         self::assertFalse($instance->hasMembers());
     }
 
@@ -71,7 +71,7 @@ final class InnerListTest extends TestCase
         $container->replace(0, ByteSequence::fromDecoded('Hello World'));
 
         self::assertCount(3, $container);
-        self::assertFalse($container->isEmpty());
+        self::assertFalse($container->hasNoMembers());
         self::assertSame('(:SGVsbG8gV29ybGQ=: 42.0 42)', $container->toHttpValue());
     }