+++ src/OrderedListTest.php
@@ -24,7 +24,7 @@ final class OrderedListTest extends StructuredFieldTest
         $instance = OrderedList::fromList($arrayParams);
 
         self::assertSame($stringItem, $instance->get(0));
-        self::assertFalse($instance->isEmpty());
+        self::assertFalse($instance->hasNoMembers());
         self::assertTrue($instance->hasMembers());
 
         self::assertEquals($arrayParams, iterator_to_array($instance));
@@ -56,7 +56,7 @@ final class OrderedListTest extends StructuredFieldTest
 
         $instance->remove(0, 1);
         self::assertCount(0, $instance);
-        self::assertTrue($instance->isEmpty());
+        self::assertTrue($instance->hasNoMembers());
     }
 
     /** @test */
@@ -69,10 +69,10 @@ final class OrderedListTest extends StructuredFieldTest
         $instance->replace(0, Item::from(ByteSequence::fromDecoded('Hello World')));
 
         self::assertCount(3, $instance);
-        self::assertFalse($instance->isEmpty());
+        self::assertFalse($instance->hasNoMembers());
         self::assertSame(':SGVsbG8gV29ybGQ=:, 42.0, 42', $instance->toHttpValue());
         $instance->clear();
-        self::assertTrue($instance->isEmpty());
+        self::assertTrue($instance->hasNoMembers());
     }
 
     /** @test */