+++ CHANGELOG.md
@@ -6,9 +6,9 @@ All Notable changes to `bakame/http-strucured-fields` will be documented in this
 
 ### Added
 
-- Introduce the`StructuredFieldContainer` interface.
-- `OrderedList`, `InnerList` now implements the `StructuredFieldList` interface.
-- `Parameters` and `Dictionnary` now implements the PHP `StructuredFieldOrderedMap` interface.
+- Introduce the`MemberContainer` interface.
+- `OrderedList`, `InnerList` now implements the `MemberList` interface.
+- `Parameters` and `Dictionnary` now implements the PHP `MemberOrderedMap` interface.
 - `Token::value` is a readonly property.
 - `Item::value` method returns the decoded value of an Item (returns value can be `float|int|string|bool`).
 - `Item::fromToken`, `Item::fromDecodedByteSequence` , `Item::fromEncodedByteSequence` to ease specific string initiation